Factor out project/schema_admin.py.
