+++ project/schema.py
@@ -4,6 +4,7 @@ from . import schema_registry
 
 # Importing this module will register our base queries/mutations.
 from . import schema_base  # noqa
+from . import schema_admin  # noqa
 
 
 SessionInfo = schema_registry.build_session_info()
