+++ texting_history/schema.py
@@ -1,18 +1,16 @@
-from typing import List, Optional, Dict, Any, Union
-import logging
-from functools import wraps
+from typing import List, Dict, Any, Union
 from pathlib import Path
 from django.db import connection
-from graphql import ResolveInfo
 import graphene
 from graphene import Mutation
-from graphene_django.types import DjangoObjectType
 
 from project import schema_registry
-from users.models import VIEW_TEXT_MESSAGE_PERMISSION, CHANGE_USER_PERMISSION, JustfixUser
-from twofactor.util import is_request_user_verified
+from project.schema_admin import (
+    ensure_request_has_verified_user_with_permission,
+    is_request_verified_user_with_permission,
+)
+from users.models import VIEW_TEXT_MESSAGE_PERMISSION
 from project.util.streaming_json import generate_json_rows
-from rapidpro.models import get_group_names_for_user
 from .management.commands.update_texting_history import update_texting_history
 from .models import Message
 from .query_parser import Query
@@ -22,8 +20,6 @@ MY_DIR = Path(__file__).parent.resolve()
 CONVERSATIONS_SQL_FILE = MY_DIR / "conversations.sql"
 
 DEFAULT_PAGE_SIZE = 50
-
-logger = logging.getLogger(__name__)
 
 
 class TextMessage(graphene.ObjectType):
@@ -59,73 +55,6 @@ class TextMessagesResult(graphene.ObjectType):
 class LatestTextMessagesResult(graphene.ObjectType):
     messages = graphene.List(graphene.NonNull(LatestTextMessage), required=True)
     has_next_page = graphene.Boolean(required=True)
-
-
-class JustfixUserType(DjangoObjectType):
-    class Meta:
-        model = JustfixUser
-        only_fields = (
-            "id",
-            "username",
-            "phone_number",
-            "first_name",
-            "last_name",
-            "email",
-            "onboarding_info",
-            "letter_request",
-        )
-
-    admin_url = graphene.String(required=True)
-
-    rapidpro_groups = graphene.Field(
-        graphene.NonNull(graphene.List(graphene.NonNull(graphene.String))),
-        description=(
-            "The RapidPro groups the user is associated with. Note that this "
-            "may be out-of-sync with the RapidPro server."
-        ),
-    )
-
-    def resolve_admin_url(self, info):
-        return self.admin_url
-
-    def resolve_rapidpro_groups(self, info) -> List[str]:
-        return get_group_names_for_user(self)
-
-
-def is_request_verified_user_with_permission(request, permission: str):
-    user = request.user
-
-    if not user.is_authenticated:
-        logger.info("User must be authenticated!")
-        return False
-
-    if not user.is_staff:
-        logger.info("User must be staff!")
-        return False
-
-    if not is_request_user_verified(request):
-        logger.info("User must be verified via two-factor authentication!")
-        return False
-
-    if not user.has_perm(permission):
-        logger.info("User does not have permission to view text messages!")
-        return False
-
-    return True
-
-
-def ensure_request_has_verified_user_with_permission(permission: str):
-    def decorator(fn):
-        @wraps(fn)
-        def wrapper(parent, info: ResolveInfo, *args, **kwargs):
-            if not is_request_verified_user_with_permission(info.context, permission):
-                return None
-
-            return fn(parent, info, *args, **kwargs)
-
-        return wrapper
-
-    return decorator
 
 
 @ensure_request_has_verified_user_with_permission(VIEW_TEXT_MESSAGE_PERMISSION)
@@ -245,51 +174,6 @@ def resolve_conversations(
         return LatestTextMessagesResult(messages=messages, has_next_page=count > len(messages))
 
 
-@ensure_request_has_verified_user_with_permission(CHANGE_USER_PERMISSION)
-def resolve_user_admin_details(
-    parent, info, phone_number: Optional[str] = None, email: Optional[str] = None
-) -> Optional[JustfixUser]:
-    if phone_number:
-        phone_number = normalize_phone_number(phone_number)
-        return JustfixUser.objects.filter(phone_number=phone_number).first()
-    elif email:
-        return JustfixUser.objects.filter(email__iexact=email).first()
-    # TODO: Maybe raise some kind of error?
-    return None
-
-
-@ensure_request_has_verified_user_with_permission(CHANGE_USER_PERMISSION)
-def resolve_user_search(parent, info, query: str) -> List[JustfixUser]:
-    from users.admin import JustfixUserAdmin
-    from project.admin import JustfixAdminSite
-
-    MAX_USER_SEARCH_RESULTS = 10
-
-    if not query:
-        return []
-
-    users, has_duplicates = JustfixUserAdmin(JustfixUser, JustfixAdminSite()).get_search_results(
-        info.context, JustfixUser.objects.all(), query
-    )
-    return list(users[:MAX_USER_SEARCH_RESULTS])
-
-
-def normalize_phone_number(phone_number: str) -> str:
-    """
-    Given either a 10-digit phone number or a U.S. phone number in E.164 format,
-    returns its 10-digit representation.
-
-    >>> normalize_phone_number('5551234567')
-    '5551234567'
-    >>> normalize_phone_number('+15551234567')
-    '5551234567'
-    """
-
-    if phone_number.startswith("+1"):
-        phone_number = phone_number[2:]
-    return phone_number
-
-
 @schema_registry.register_queries
 class TextingHistory:
     conversations = graphene.Field(
@@ -306,25 +190,6 @@ class TextingHistory:
         first=graphene.Int(default_value=DEFAULT_PAGE_SIZE),
         after_or_at=graphene.Float(default_value=0),
         resolver=resolve_conversation,
-    )
-
-    user_details = graphene.Field(
-        JustfixUserType,
-        phone_number=graphene.String(),
-        email=graphene.String(),
-        resolver=resolve_user_admin_details,
-    )
-
-    user_search = graphene.Field(
-        graphene.List(graphene.NonNull(JustfixUserType)),
-        query=graphene.String(required=True),
-        resolver=resolve_user_search,
-    )
-
-    is_verified_staff_user = graphene.Boolean(
-        resolver=ensure_request_has_verified_user_with_permission(CHANGE_USER_PERMISSION)(
-            lambda parent, info: True
-        ),
     )
